pyproject.toml
[project]
name = "bot-service"
version = "0.1.0"
description = "Telegram bot and Flask-based RAG editor service."
authors = [{ name = "Denis Varekha", email = "d.varekha@gmail.com" }]
requires-python = ">=3.12,<4.0"

dependencies = [
    "python-telegram-bot[job-queue]",
    "python-dotenv",
    "openai==0.28.0",
    # "sqlalchemy>=2.0,<3.0",
]

[dependency-groups]
dev = [
    "ruff==0.11.13",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion    = "3.12"
venvPath = "."
venv     = ".venv"
include = ["telegram_bot", "tests"]
